requirements.txt
# Core data science
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
lightgbm==4.1.0

# Visualization
matplotlib==3.8.2
seaborn==0.13.0

# API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Serialization
joblib==1.3.2
cloudpickle==3.0.0

# Experiment tracking
mlflow==2.8.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# HTTP requests (для тестирования API)
requests==2.31.0

# Utilities
tqdm==4.66.1
typing-extensions==4.8.0